be explicit that this change is to allow the same pattern across 2 clients
